refactor: 로그인 관련 클래스를 global/login으로 이동
