+++ server/src/main/java/com/onetool/server/counting/ServiceCountingService.java
@@ -2,8 +2,7 @@ package com.onetool.server.counting;
 
 import com.onetool.server.blueprint.BlueprintRepository;
 import com.onetool.server.counting.dto.ServiceCountingResponse;
-import com.onetool.server.member.MemberRepository;
-import lombok.NoArgsConstructor;
+import com.onetool.server.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
