+++ server/src/main/java/com/onetool/server/global/auth/Oauth.java
@@ -1,6 +1,6 @@
 package com.onetool.server.global.auth;
 
-import com.onetool.server.member.Member;
+import com.onetool.server.member.domain.Member;
 import jakarta.annotation.Nullable;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
