+++ server/src/main/java/com/onetool/server/global/auth/login/userinfo/GoogleOAuth2UserInfo.java
@@ -1,4 +1,4 @@
-package com.onetool.server.global.oauth2.userinfo;
+package com.onetool.server.global.auth.login.userinfo;
 
 import java.util.Map;
