+++ server/src/main/java/com/onetool/server/member/controller/SSOTestController.java
@@ -1,4 +1,4 @@
-package com.onetool.server.member;
+package com.onetool.server.member.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
