+++ server/src/main/java/com/onetool/server/member/enums/SocialType.java
@@ -1,4 +1,4 @@
-package com.onetool.server.member;
+package com.onetool.server.member.enums;
 
 public enum SocialType {
     NAVER, KAKAO, GOOGLE
