+++ server/src/main/java/com/onetool/server/member/enums/UserRole.java
@@ -1,4 +1,4 @@
-package com.onetool.server.member;
+package com.onetool.server.member.enums;
 
 public enum UserRole {
     ROLE_USER, ROLE_ADMIN, ROLE_GUEST
