+++ server/src/main/java/com/onetool/server/member/service/MemberService.java
@@ -1,8 +1,10 @@
-package com.onetool.server.member;
+package com.onetool.server.member.service;
 
 import com.onetool.server.global.auth.MemberAuthContext;
 import com.onetool.server.global.auth.jwt.JwtUtil;
 import com.onetool.server.global.exception.MemberNotFoundException;
+import com.onetool.server.member.repository.MemberRepository;
+import com.onetool.server.member.domain.Member;
 import com.onetool.server.member.dto.LoginRequest;
 import com.onetool.server.member.dto.MemberCreateRequest;
 import com.onetool.server.member.dto.MemberCreateResponse;
@@ -11,7 +13,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
